fix scenario security rules looping issue
There's a bug in the `postgrespy`'s ArrayField. The ArrayField object does not reset the index every time it is iterated. And the `rq` seems to iterate the `scenario.sg_rules` when passing it through the redis, thus preventing the following functions from iterating it.

Here is a proof-of-concept example:

```py
import fields
x = fields.ArrayField([1,2,3])
for i in x:
    print(i)

# 1
# 2
# 3

for i in x:
    print(i)
# nothing
```